+++ ReleaseNotes.md
@@ -7,3 +7,7 @@
 - Fixes small bugs in NuGet packaging
 - Adds fluent API to `AppInfoManager`/`AppInfoHandler`
 - Updates to documentation
+
+# 0.1.2
+- Adds new `DoNotBuild()` method to prevent building the app before publishing.
+- Fixes the `WithVersion` extension to add ".*" segments as needed.